add testing logger
